+++ torchkge/exceptions/__init__.py
@@ -1,0 +1,3 @@
+from .exceptions import NotYetEvaluatedError
+from .exceptions import SizeMismatchError
+from .exceptions import WrongDimensionError